Refactor class logic to use base classes
